% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adas_class}
\alias{adas_class}
\title{Some utility functions, mainly used for \code{add_ad_class}; also contains \code{adas_plotter} call}
\usage{
adas_class(adas_curr)
}
\description{
Some utility functions, mainly used for \code{add_ad_class}; also contains \code{adas_plotter} call
}
