% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adas_plotter.R
\name{adas_plotter}
\alias{adas_plotter}
\title{A function to plot the beginning and ending classifications}
\usage{
adas_plotter(adas_final)
}
\description{
@param adas_final A dataframe of ADAS-Cog scores and classifications
}
\details{
@export

@return A list of plots, 'bl' for baseline and 'end' for last visit
}
