% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ad_class.R
\name{add_ad_class}
\alias{add_ad_class}
\title{Take the simulated ADAS-Cog scores and probablistically get a cognitive status}
\usage{
add_ad_class(adas_calc)
}
\arguments{
\item{adas_calc}{A dataframe with ADAS-Cog scores (named "Adas")}
}
\value{
The dataframe with classifications included
}
\description{
See \code{utils.r} for helper function used to calculate probabilities based on ADAS-Cog score
}
