% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_placebo_vals.R
\name{add_placebo_vals}
\alias{add_placebo_vals}
\title{Takes the subject specific ADAS-Cog scores and add placebo modifications, donepezil use, and random noise}
\usage{
add_placebo_vals(adas_theta, pl_blocks)
}
\arguments{
\item{adas_theta}{The list of dataframes of longitudinal logit-scale ADAS-Cog scores with covariates}

\item{pl_blocks}{The list of median parameter values made by \code{make_posterior_blocks}}
}
\value{
A final dataframe combining all blocks and indexed by study and subject with placebo modifications
}
\description{
These modifications add a slight dip in ADAS-Cog scores early on before natural progression begins again.
The donepezil use is based on 5mg usage from adsim.  Finally, random noise follows a normal distribution.
}
