% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_posterior_blocks.R
\name{make_posterior_blocks}
\alias{make_posterior_blocks}
\title{Create the sample blocks from a posterior matrix}
\usage{
make_posterior_blocks(posterior, n = 50, blocks = 10)
}
\arguments{
\item{posterior}{Matrix of posterior parameters from the adsim package}

\item{n}{The number of samples per block or a vector of length blocks (default 50)}

\item{blocks}{the number of blocks (default 10)}
}
\value{
A list of single vectors (median parameter values) blocks
}
\description{
Create the sample blocks from a posterior matrix
}
