% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_subject_vars.R
\name{set_subject_vars}
\alias{set_subject_vars}
\title{Set the subject-specific random intercepts and slopes for}
\usage{
set_subject_vars(pl_blocks)
}
\arguments{
\item{popPars_curr}{The list of sampled population parameters taken from \code{make_posterior_blocks}}
}
\value{
Another list with the subject-specific values
}
\description{
Set the subject-specific random intercepts and slopes for
}
