% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_times.R
\name{set_times}
\alias{set_times}
\title{Add times to the baseline values and calculate the subject-specific ADAS-Cog scores on the logit scale}
\usage{
set_times(bl_vars)
}
\value{
A list of studies which each has a long dataframe with the sample times and calculated ADAS-Cog scores
}
\description{
Add times to the baseline values and calculate the subject-specific ADAS-Cog scores on the logit scale
}
